backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.6.1
python-dotenv==1.0.1
typing-extensions==4.10.0
